+++ .rubocop.yml
@@ -1,4 +1,4 @@
-`AllCops:
+AllCops:
   NewCops: enable
   Exclude:
     - "Guardfile"
